backend/requirements.txt
requests==2.31.0
numpy==1.26.4
python-dotenv==1.0.1
flask==3.0.0
flask-cors==4.0.0
